Moved the CachedObjects to webide package
