+++ modules/admin/app/views/admin/index.scala.html
@@ -1,8 +1,3 @@
-@import java.util
-@import models.common.database.Project
-
-@(projects: util.List[Project], defaultProject: Project)
-
-@views.html.common.main("RESOLVEWebIDE - Admin", projects, defaultProject, false) {
+@views.html.common.main("RESOLVEWebIDE - Admin") {
     <div>Admin Page</div>
 }
