+++ modules/common/app/views/common/authentication/registration.scala.html
@@ -1,9 +1,4 @@
-@import java.util
-@import models.common.database.Project
-
-@(projects: util.List[Project], activeProject: Project)
-
-@views.html.common.main("RESOLVEWebIDE - Registration", projects, activeProject, false) {
+@views.html.common.main("RESOLVEWebIDE - Registration") {
     <div class="container" style="padding-top:30px; padding-bottom:30px;">
         <div class="panel panel-primary">
             <div class="panel-heading">
