+++ modules/common/app/views/common/tags/component_bar.scala.html
@@ -1,8 +1,35 @@
-<div class="btn-toolbar" role="toolbar" aria-label="components-toolbar">
-    <!--<div class="btn-group" role="group" aria-label="...">
-        <a href="#" class="btn btn-primary" role="button">Projects</a>
-    </div>
-    <div class="btn-group" role="group" aria-label="...">
-        <a href="#" class="btn btn-primary" role="button">Components</a>
-    </div>-->
-</div>
+@import java.util
+@import models.common.database.Project
+
+@(projects: util.List[Project], activeProject: Project)
+
+<nav class="navbar navbar-default navbar-fixed-top" aria-label="components-toolbar">
+    <ul class="nav navbar-nav">
+        <li class="dropdown">
+            <div class="btn-group">
+                <button type="button" class="btn btn-default navbar-btn" style="margin-bottom: 0px ;">
+                    <a href="?p=@activeProject.id">@activeProject.name</a>
+                </button>
+                <button type="button" class="btn btn-default navbar-btn dropdown-toggle" style="margin-bottom: 0px ;" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+                    <span class="caret"></span>
+                    <span class="sr-only">Toggle Dropdown</span>
+                </button>
+                <ul class="dropdown-menu">
+                @if(projects.size == 1) {
+                    <li><p></p></li>
+                    <li><p class="text-center">None Available</p></li>
+                } else {
+                    @for(p <- projects) {
+                        @if(p.name != activeProject.name) {
+                            <li><a href="?p=@p.id">@p.name</a></li>
+                        }
+                    }
+                }
+                </ul>
+            </div>
+        </li>
+        <button type="button" class="btn btn-default navbar-btn" style="margin-left: 10px ; margin-bottom: 0px ;">
+            <a href="#">Components</a>
+        </button>
+    </ul>
+</nav>
