+++ modules/webide/app/controllers/webide/CachedObjects.java
@@ -1,4 +1,4 @@
-package controllers.common.utilities;
+package controllers.webide;
 
 import models.common.database.Project;
 import java.util.List;
