+++ modules/webide/app/views/webide/index.scala.html
@@ -3,9 +3,9 @@
 
 @(projects: util.List[Project], activeProject: Project)
 
-@views.html.common.main("RESOLVEWebIDE", projects, activeProject, true) {
+@views.html.common.main("RESOLVEWebIDE") {
     <!-- Code to display components bar -->
-    @views.html.common.tags.component_bar()
+    @views.html.common.tags.component_bar(projects, activeProject)
 
     <!-- Code to display ace editor -->
     <div id="editor">function foo(items) {
